+++ Nimble/objc/DSL.m
@@ -84,3 +84,14 @@ NIMBLE_EXPORT NMBObjCRaiseExceptionMatcher *NMB_raiseException() {
     return [NMBObjCMatcher raiseExceptionMatcher];
 }
 
+NIMBLE_EXPORT NMBWaitUntilTimeoutBlock nmb_wait_until_timeout_builder(NSString *file, NSUInteger line) {
+    return ^(NSTimeInterval timeout, void (^action)(void (^)(void))) {
+        [NMBWait untilTimeout:timeout action:action file:file line:line];
+    };
+}
+
+NIMBLE_EXPORT NMBWaitUntilBlock nmb_wait_until_builder(NSString *file, NSUInteger line) {
+  return ^(void (^action)(void (^)(void))) {
+    [NMBWait until:action file:file line:line];
+  };
+}
